feature: create custom hook for Expenses and createExpense function
